Moved TableExtension to pagination package
