+++ src/main/kotlin/test/ktortemplate/core/persistance/sql/CarRepositoryImpl.kt
@@ -10,8 +10,8 @@ import test.ktortemplate.conf.database.DatabaseConnection
 import test.ktortemplate.core.model.Car
 import test.ktortemplate.core.model.CarSaveCommand
 import test.ktortemplate.core.persistance.CarRepository
-import test.ktortemplate.core.utils.createFromFilters
-import test.ktortemplate.core.utils.createSorts
+import test.ktortemplate.core.utils.pagination.createFromFilters
+import test.ktortemplate.core.utils.pagination.createSorts
 import test.ktortemplate.core.utils.pagination.PageRequest
 
 class CarRepositoryImpl : CarRepository, KoinComponent {
