+++ src/main/kotlin/test/ktortemplate/core/utils/pagination/TableExtensions.kt
@@ -1,4 +1,4 @@
-package test.ktortemplate.core.utils
+package test.ktortemplate.core.utils.pagination
 
 import org.jetbrains.exposed.sql.AutoIncColumnType
 import org.jetbrains.exposed.sql.BooleanColumnType
@@ -16,9 +16,6 @@ import org.jetbrains.exposed.sql.compoundAnd
 import org.jetbrains.exposed.sql.compoundOr
 import org.jetbrains.exposed.sql.select
 import org.jetbrains.exposed.sql.selectAll
-import test.ktortemplate.core.utils.pagination.FilterField
-import test.ktortemplate.core.utils.pagination.SortField
-import test.ktortemplate.core.utils.pagination.SortFieldOrder
 
 /**
  * Creates exposed query from list of FilterField
